+++ packages/amplify-util-mock/src/__e2e__/connections-with-auth-tests.e2e.test.ts
@@ -131,7 +131,7 @@ type Stage @model @auth(rules: [{ allow: groups, groups: ["Admin"]}]) {
     // Verify we have all the details
     expect(GRAPHQL_ENDPOINT).toBeTruthy();
 
-    const idToken = signUpAddToGroupAndGetJwtToken(USER_POOL_ID, USERNAME1, USERNAME1, [
+    const idToken = await signUpAddToGroupAndGetJwtToken(USER_POOL_ID, USERNAME1, USERNAME1, [
       ADMIN_GROUP_NAME,
       WATCHER_GROUP_NAME,
       PARTICIPANT_GROUP_NAME,
@@ -140,12 +140,12 @@ type Stage @model @auth(rules: [{ allow: groups, groups: ["Admin"]}]) {
       Authorization: idToken,
     });
 
-    const idToken2 = signUpAddToGroupAndGetJwtToken(USER_POOL_ID, USERNAME2, USERNAME2, [DEVS_GROUP_NAME]);
+    const idToken2 = await signUpAddToGroupAndGetJwtToken(USER_POOL_ID, USERNAME2, USERNAME2, [DEVS_GROUP_NAME]);
     GRAPHQL_CLIENT_2 = new GraphQLClient(GRAPHQL_ENDPOINT, {
       Authorization: idToken2,
     });
 
-    const idToken3 = signUpAddToGroupAndGetJwtToken(USER_POOL_ID, USERNAME3, USERNAME3, []);
+    const idToken3 = await signUpAddToGroupAndGetJwtToken(USER_POOL_ID, USERNAME3, USERNAME3, []);
     GRAPHQL_CLIENT_3 = new GraphQLClient(GRAPHQL_ENDPOINT, {
       Authorization: idToken3,
     });