+++ packages/amplify-java-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.39](https://github.com/aws-amplify/amplify-cli/compare/amplify-java-function-runtime-provider@2.3.38...amplify-java-function-runtime-provider@2.3.39) (2024-03-29)
+
+**Note:** Version bump only for package amplify-java-function-runtime-provider
+
+
+
+
+
 ## [2.3.38](https://github.com/aws-amplify/amplify-cli/compare/amplify-java-function-runtime-provider@2.3.37...amplify-java-function-runtime-provider@2.3.38) (2024-03-06)
 
 **Note:** Version bump only for package amplify-java-function-runtime-provider